+++ Kexa/main.ts
@@ -46,15 +46,13 @@ export async function main() {
     let settings = await gatheringRules(await getEnvVar("RULESDIRECTORY")??"./Kexa/rules");
     context?.log("settings", settings);
     if(settings.length != 0){
-        let resources = {};
-        resources = await loadAddOns(resources);
+        let resources = await loadAddOns(settings);
         context?.log("resources", resources);
         if(args.o) createFileSync(JSON.stringify(resources), folderOutput + "/resources/"+ new Date().toISOString().slice(0, 16).replace(/[-T:/]/g, '') +".json", true);
         context?.log("good");
         settings.forEach(setting => {
             context?.log("setting", setting);
             let result = checkRules(setting.rules, resources, setting.alert);
-            logger.debug(result);
             if(setting.alert.global.enabled){
                 let render_table = renderTableAllScan(result.map(scan => scan.filter(value => value.error.length>0)));
                 let render_table_loud = renderTableAllScanLoud(result.map(scan => scan.filter(value => value.loud)));