+++ Kexa/services/analyse.service.ts
@@ -22,6 +22,7 @@ import { extractHeaders } from './addOn.service';
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 import {getContext, getNewLogger} from "./logger.service";
+import { splitProperty } from '../helpers/spliter';
 const logger = getNewLogger("AnalyseLogger");
 
 const jsome = require('jsome');
@@ -33,7 +34,7 @@ const varEnvMin = {
 const config = require('node-config-ts').config;
 const levelAlert = ["info", "warning", "error", "critical"];
 let headers: any;
-//Analyse  list
+// Analyze  list
 // read the yaml file with rules
 // exam each rules and raise alarm or not
 export async function gatheringRules(rulesDirectory:string, getAll:boolean=false): Promise<SettingFile[]> {
@@ -46,7 +47,7 @@ export async function gatheringRules(rulesDirectory:string, getAll:boolean=false
     let listNeedRules = getListNeedRules();
     for(const p of paths) {
         logger.debug("getting "+rulesDirectory+"/"+p.name+" rules.");
-        let setting = await analyseRule(rulesDirectory+"/"+p.name, listNeedRules, getAll);
+        let setting = await analyzeRule(rulesDirectory+"/"+p.name, listNeedRules, getAll);
         if(setting){
             setting.alert.global.name = p.name.split(".")[0];
             settingFileList.push(setting);
@@ -60,32 +61,41 @@ export async function gatheringRules(rulesDirectory:string, getAll:boolean=false
 
 export function extractAddOnNeed(settingFileList: SettingFile[]){
     let providerList = new Array<string>();
+    let objectNameList:any = {};
     settingFileList.forEach((ruleFile) => {
+        objectNameList[ruleFile.alert.global.name] = {};
         ruleFile.rules.forEach((rule) => {
             if(!providerList.includes(rule.cloudProvider)) providerList.push(rule.cloudProvider);
+            if(!objectNameList[ruleFile.alert.global.name][rule.cloudProvider]) objectNameList[ruleFile.alert.global.name][rule.cloudProvider] = new Array<string>();
+            if(!objectNameList[ruleFile.alert.global.name][rule.cloudProvider].includes(rule.objectName)) objectNameList[ruleFile.alert.global.name][rule.cloudProvider].push(rule.objectName);
         });
     });
-    writeStringToJsonFile(JSON.stringify({ "addOn" : providerList}), "./config/addOnNeed.json");
+    writeStringToJsonFile(JSON.stringify({ "addOn" : providerList, "objectNameNeed": objectNameList }), "./config/addOnNeed.json");
 }
 
 function getListNeedRules(): string[]{
     const config = require('node-config-ts').config;
     let listNeedRules = new Array<string>();
     for(let cloudProvider of Object.keys(config)){
+        if(["host", "host", "workerId", "requestId", "grpcMaxMessageLength"].includes(cloudProvider)) continue;
         let configAssign = config[cloudProvider];
-        for(let config of configAssign){
-            if (Array.isArray(config.rules)) {
-                for (let rule of config.rules) {
-                    if (!listNeedRules.includes(rule)) listNeedRules.push(rule);
+        try{
+            for(let config of configAssign){
+                if (Array.isArray(config.rules)) {
+                    for (let rule of config.rules) {
+                        if (!listNeedRules.includes(rule)) listNeedRules.push(rule);
+                    }
                 }
             }
+        }catch(err){
+            logger.debug("error in getListNeedRules:"+err);
         }
     }
     return listNeedRules;
 }
 
-export async function analyseRule(ruleFilePath:string, listNeedRules:string[], getAll:boolean=false): Promise<SettingFile | null> {
-    logger.debug("analyse:"+ruleFilePath);
+export async function analyzeRule(ruleFilePath:string, listNeedRules:string[], getAll:boolean=false): Promise<SettingFile | null> {
+    logger.debug("analyze:"+ruleFilePath);
     try {
         let lastBlockSplited = ruleFilePath.split('/')[ruleFilePath.split('/').length -1].split(".");
         if(lastBlockSplited.length == 1){
@@ -125,9 +135,9 @@ export async function checkDoc(doc:SettingFile): Promise<string[]> {
     logger.debug("check doc");
     let result:string[] = [];
     if(!doc.hasOwnProperty("version")) result.push("info - version not found in doc");
-    else if(doc.version.match(/^[0-9]+\.[0-9]+\.[0-9]+$/) === null) result.push("debug - version not valid in doc : "+ doc.version);
+    else if(RegExp(/^[0-9]+\.[0-9]+\.[0-9]+$/).exec(doc.version) === null) result.push("debug - version not valid in doc : "+ doc.version);
     if(!doc.hasOwnProperty("date")) result.push("info - date not found in doc");
-    else if(doc.date.match(/^(0[1-9]|[12][0-9]|3[01])-(0[1-9]|1[012])-(19|20)\d\d$/) === null) result.push("debug - date not valid in doc : "+ doc.date);
+    else if(RegExp(/^(0[1-9]|[12][0-9]|3[01])-(0[1-9]|1[012])-(19|20)\d\d$/).exec(doc.date) === null) result.push("debug - date not valid in doc : "+ doc.date);
     (await checkDocAlert(doc.alert)).forEach((value) => result.push(value));
     checkDocRules(doc.rules).forEach((value) => result.push(value));
     return result;
@@ -421,7 +431,7 @@ export function parentResultScan(subResultScans: SubResultScan[], result: boolea
         value: null,
         condition: subResultScans.map((value) => value.condition).flat(),
         result,
-        message : subResultScans.map((value) => value.message).join(" || ")
+        message : subResultScans.map((value) => value.message).filter(item => item != "").join(" || ")
     };
 }
 
@@ -521,7 +531,7 @@ export function resultScan(condition: RulesConditions, value: any, fs: Function[
 
 export function getSubProperty(object:any, property:string): any {
     if (property === ".")  return object;
-    let properties = property.split(".");
+    let properties = splitProperty(property, ".", "/");
     let result = object;
     properties.forEach(prop => {
         result = result[prop];